Create  leapyear.java
The code checks these rules to see if a year is a leap year:

If the year can be divided evenly by 4, go to the next step. Otherwise, it's not a leap year.
If the year can be divided evenly by 100, go to the next step. Otherwise, it is a leap year.
If the year can be divided evenly by 400, then it is a leap year. Otherwise, it's not a leap year.